tests: versions of parsers
