+++ test/wos/test_wos_parser_v_1_0_0.py
@@ -1,5 +1,5 @@
 #!/usr/bin/env python3
-"""Web-of-Science query parser unit tests."""
+"""Tests for WOSParser_v1_0_0"""
 import typing
 
 import pytest
@@ -15,8 +15,8 @@ from search_query.query import SearchField
 from search_query.query_and import AndQuery
 from search_query.query_or import OrQuery
 from search_query.query_term import Term
-from search_query.wos.parser import WOSListParser
-from search_query.wos.parser import WOSParser
+from search_query.wos.v1_0_0.parser import WOSListParser_v1_0_0
+from search_query.wos.v1_0_0.parser import WOSParser_v1_0_0
 
 # ruff: noqa: E501
 # flake8: noqa: E501
@@ -88,7 +88,7 @@ from search_query.wos.parser import WOSParser
     ],
 )
 def test_tokenization(query_str: str, expected_tokens: list) -> None:
-    parser = WOSParser(query_str=query_str)
+    parser = WOSParser_v1_0_0(query_str=query_str)
     parser.tokenize()
     assert parser.tokens == expected_tokens, print(parser.tokens)
 
@@ -737,7 +737,7 @@ def test_linter(
     expected_messages: typing.List[dict],
 ) -> None:
     print(query_str)
-    parser = WOSParser(query_str)
+    parser = WOSParser_v1_0_0(query_str)
     try:
         parser.parse()
     except SearchQueryException:
@@ -774,7 +774,7 @@ def test_linter(
 )
 def test_implicit_precedence(query_str: str, expected_query: str) -> None:
     print(query_str)
-    parser = WOSParser(query_str)
+    parser = WOSParser_v1_0_0(query_str)
     query = parser.parse()
     parser.print_tokens()
 
@@ -792,7 +792,7 @@ def test_implicit_precedence(query_str: str, expected_query: str) -> None:
 
 def test_query_parsing_basic_vs_advanced() -> None:
     # Basic search
-    parser = WOSParser(query_str="digital AND online")
+    parser = WOSParser_v1_0_0(query_str="digital AND online")
     parser.parse()
     assert parser.linter.messages == [
         {
@@ -806,34 +806,36 @@ def test_query_parsing_basic_vs_advanced() -> None:
     ]
 
     # Search field could be nested
-    parser = WOSParser(
+    parser = WOSParser_v1_0_0(
         query_str="(TI=digital AND AB=online)",
     )
     parser.parse()
     assert len(parser.linter.messages) == 0
 
     # Advanced search
-    parser = WOSParser(query_str="ALL=(digital AND online)")
+    parser = WOSParser_v1_0_0(query_str="ALL=(digital AND online)")
     parser.parse()
     assert len(parser.linter.messages) == 0
 
-    parser = WOSParser(query_str="(ALL=digital AND ALL=online)")
+    parser = WOSParser_v1_0_0(query_str="(ALL=digital AND ALL=online)")
     parser.parse()
     assert len(parser.linter.messages) == 0
 
     # ERROR: Basic search without field_general
-    parser = WOSParser(query_str="digital AND online")
+    parser = WOSParser_v1_0_0(query_str="digital AND online")
     parser.parse()
     assert len(parser.linter.messages) == 1
 
     # ERROR: Advanced search with field_general
-    parser = WOSParser(query_str="ALL=(digital AND online)", field_general="All Fields")
+    parser = WOSParser_v1_0_0(
+        query_str="ALL=(digital AND online)", field_general="All Fields"
+    )
     parser.parse()
     assert len(parser.linter.messages) == 1
     print(parser.linter.messages)
 
     # ERROR: Advanced search with field_general
-    parser = WOSParser(query_str="TI=(digital AND online)")
+    parser = WOSParser_v1_0_0(query_str="TI=(digital AND online)")
     parser.parse()
     assert len(parser.linter.messages) == 0
 
@@ -865,7 +867,7 @@ def test_query_parsing_basic_vs_advanced() -> None:
 )
 def test_parser_wos(query_str: str, expected_translation: str) -> None:
     print(query_str)
-    wos_parser = WOSParser(query_str)
+    wos_parser = WOSParser_v1_0_0(query_str)
     query_tree = wos_parser.parse()
     assert expected_translation == query_tree.to_generic_string(), print(
         query_tree.to_generic_string()
@@ -873,7 +875,7 @@ def test_parser_wos(query_str: str, expected_translation: str) -> None:
 
 
 def test_query_in_quotes() -> None:
-    parser = WOSParser(query_str='"TI=(digital AND online)"')
+    parser = WOSParser_v1_0_0(query_str='"TI=(digital AND online)"')
     parser.parse()
 
     # Assertions using standard assert statement
@@ -882,7 +884,7 @@ def test_query_in_quotes() -> None:
 
 
 def test_artificial_parentheses() -> None:
-    parser = WOSParser(
+    parser = WOSParser_v1_0_0(
         query_str="remote OR online AND work",
         field_general="All Fields",
     )
@@ -930,7 +932,7 @@ def test_artificial_parentheses() -> None:
 def test_list_parser_case_1() -> None:
     query_list = '1. TS=("Peer leader*" OR "Shared leader*" OR "Distributed leader*" OR "Distributive leader*" OR "Collaborate leader*" OR "Collaborative leader*" OR "Team leader*" OR "Peer-led" OR "Athlete leader*" OR "Team captain*" OR "Peer mentor*" OR "Peer Coach")\n2. TS=("acrobatics" OR "acrobat" OR "acrobats" OR "acrobatic" OR "aikido" OR "aikidoists" OR "anetso" OR "archer" OR "archers" OR "archery" OR "airsoft" OR "angling" OR "aquatics" OR "aerobics" OR "athlete" OR "athletes" OR "athletic" OR "athletics" OR "ball game*" OR "ballooning" OR "basque pelota" OR "behcup" OR "bicycling" OR "BMX" OR "bodyboarding" OR "boule lyonnaise" OR "bridge" OR "badminton" OR "balle au tamis" OR "baseball" OR "basketball" OR "battle ball" OR "battleball" OR "biathlon" OR "billiards" OR "boating" OR "bobsledding" OR "bobsled" OR "bobsledder" OR "bobsledders" OR "bobsleigh" OR "boccia" OR "bocce" OR "buzkashi" OR "bodybuilding" OR "bodybuilder" OR "bodybuilders" OR "bowling" OR "bowler" OR "bowlers" OR "bowls" OR "boxing" OR "boxer" OR "boxers" OR "bandy" OR "breaking" OR "breakdanc*" OR "broomball" OR "budo" OR "bullfighting" OR "bullfights" OR "bullfight" OR "bullfighter" OR "bullfighters" OR "mountain biking" OR "mountain bike" OR "carom billiards" OR "camogie" OR "canoe slalom" OR "canoeing" OR "canoeist" OR "canoeists" OR "canoe" OR "climbing" OR "coasting" OR "cricket" OR "croquet" OR "crossfit" OR "curling" OR "curlers" OR "curler" OR "cyclist" OR "cyclists" OR "combat*" OR "casting" OR "cheerleading" OR "cheer" OR "cheerleader*" OR "chess" OR "charrerias" OR "cycling" OR "dancesport" OR "darts" OR "decathlon" OR "draughts" OR "dancing" OR "dance" OR "dancers" OR "dancer" OR "diving" OR "dodgeball" OR "e-sport" OR "dressage" OR "endurance" OR "equestrian" OR "eventing" OR "eskrima" OR "escrima" OR "fencer" OR "fencing" OR "fencers" OR "fishing" OR "finswimming" OR "fistball" OR "floorball" OR "flying disc" OR "foosball" OR "futsal" OR "flickerball" OR "football" OR "frisbee" OR "gliding" OR "go" OR "gongfu" OR "gong fu" OR "goalball" OR "golf" OR "golfer" OR "golfers" OR "gymnast" OR "gymnasts" OR "gymnastics" OR "gymnastic" OR "gymkhanas" OR "half rubber" OR "highland games" OR "hap ki do" OR "halfrubber" OR "handball" OR "handballers" OR "handballer" OR "hapkido" OR "hiking" OR "hockey" OR "hsing-I" OR "hurling" OR "Hwa rang do" OR "hwarangdo" OR "horsemanship" OR "horseshoes" OR "orienteer" OR "orienteers" OR "orienteering" OR "iaido" OR "iceboating" OR "icestock" OR "intercrosse" OR "jousting" OR "jai alai" OR "jeet kune do" OR "jianzi" OR "jiu-jitsu" OR "jujutsu" OR "ju-jitsu" OR "kung fu" OR "kungfu" OR "kenpo" OR "judo" OR "judoka" OR "judoists" OR "judoist" OR "jump" OR "jumping" OR "jumper" OR "jian zi" OR "kabaddi" OR "kajukenbo" OR "karate" OR "karateists" OR "karateist" OR "karateka" OR "kayaking" OR "kendo" OR "kenjutsu" OR "kickball" OR "kickbox*" OR "kneeboarding" OR "krav maga" OR "kuk sool won" OR "kun-tao" OR "kuntao" OR "kyudo" OR "korfball" OR "lacrosse" OR "life saving" OR "lapta" OR "lawn tempest" OR "bowling" OR "bowls" OR "logrolling" OR "luge" OR "marathon" OR "marathons" OR "marathoning" OR "martial art" OR "martial arts" OR "martial artist" OR "martial artists" OR "motorsports" OR "mountainboarding" OR "mountain boarding" OR "mountaineer" OR "mountaineering" OR "mountaineers" OR "muay thai" OR "mallakhamb" OR "motorcross" OR "modern arnis" OR "naginata do" OR "netball" OR "ninepins" OR "nine-pins" OR "nordic combined" OR "nunchaku" OR "olympic*" OR "pes\u00e4pallo" OR "pitch and putt" OR "pool" OR "pato" OR "paddleball" OR "paddleboarding" OR "pankration" OR "pancratium" OR "parachuting" OR "paragliding" OR "paramotoring" OR "paraski" OR "paraskiing" OR "paraskier" OR "paraskier" OR "parakour" OR "pelota" OR "pencak silat" OR "pentathlon" OR "p\u00e9tanque" OR "petanque" OR "pickleball" OR "pilota" OR "pole bending" OR "pole vault" OR "polo" OR "polocrosse" OR "powerlifting" OR "player*" OR "powerboating" OR "pegging" OR "parathletic" OR "parathletics" OR "parasport*" OR "paraathletes" OR "paraathlete" OR "pushball" OR "push ball" OR "quidditch" OR "races" OR "race" OR "racing" OR "racewalking" OR "racewalker" OR "racewalkers" OR "rackets" OR "racketlon" OR "racquetball" OR "racquet" OR "racquets" OR "rafting" OR "regattas" OR "riding" OR "ringette" OR "rock-it-ball" OR "rogaining" OR "rock climbing" OR "roll ball" OR "roller derby" OR "roping" OR "rodeos" OR "rodeo" OR "riding" OR "rider" OR "riders" OR "rounders" OR "rowing" OR "rower" OR "rowers" OR "rug ball" OR "running" OR "runner" OR "runners" OR "rugby" OR "sailing" OR "san shou" OR "sepaktakraw" OR "sepak takraw" OR "san-jitsu" OR "savate" OR "shinty" OR "shishimai" OR "shooting" OR "singlestick" OR "single stick" OR "skateboarding" OR "skateboarder" OR "skateboarders" OR "skater" OR "skaters" OR "skating" OR "skipping" OR "racket game*" OR "rollerskating" OR "skelton" OR "skibobbing" OR "ski" OR "skiing" OR "skier" OR "skiers" OR "skydive" OR "skydiving" OR "skydivers" OR "skydiver" OR "skysurfing" OR "sledding" OR "sledging" OR "sled dog" OR "sleddog" OR "snooker" OR "sleighing" OR "snowboarder" OR "snowboarding" OR "snowboarders" OR "snowshoeing" OR "soccer" OR "softball" OR "spear fighting" OR "speed-a-way" OR "speedball" OR "sprint" OR "sprinting" OR "sprints" OR "squash" OR "stick fighting" OR "stickball" OR "stoolball" OR "stunt flying" OR "sumo" OR "surfing" OR "surfer" OR "surfers" OR "swimnastics" OR "swimming" OR "snowmobiling" OR "swim" OR "swimmer" OR "swimmers" OR "shot-put" OR "shot-putters" OR "shot-putter" OR "sport" OR "sports" OR "tae kwon do" OR "taekwondo" OR "taekgyeon" OR "taekkyeon" OR "taekkyon" OR "taekyun" OR "tang soo do" OR "tchoukball" OR "tennis" OR "tetherball" OR "throwing" OR "thrower" OR "throwers" OR "tai ji" OR "tai chi" OR "taiji" OR "t ai chi" OR "throwball" OR "tug of war" OR "tobogganing" OR "track and field" OR "track & field" OR "trampoline" OR "trampolining" OR "trampolinists" OR "trampolinist" OR "trapball" OR "trapshooting" OR "triathlon" OR "triathlete" OR "triathletes" OR "tubing" OR "tumbling" OR "vaulting" OR "volleyball" OR "wakeboarding" OR "wallyball" OR "weightlifting" OR "weightlifter" OR "weightlifters" OR "wiffle ball" OR "windsurfing" OR "windsurfer" OR "windsurfers" OR "walking" OR "wingwalking" OR "woodchopping" OR "wood chopping" OR "woodball" OR "wushu" OR "weight lifter" OR "weight lift" OR "weight lifters" OR "wrestling" OR "wrestler" OR "wrestlers" OR "vovinam" OR "vx" OR "yoga")\n3. #1 AND #2\n'
 
-    list_parser = WOSListParser(query_list=query_list)
+    list_parser = WOSListParser_v1_0_0(query_list=query_list)
     list_parser.parse()
 
 
@@ -938,7 +940,7 @@ def test_list_parser_case_1() -> None:
 def test_list_parser_case_2() -> None:
     query_list = '1. TS=("Peer leader*" OR "Shared leader*")\n2. TS=("acrobatics" OR "acrobat" OR "acrobats")'
 
-    list_parser = WOSListParser(query_list=query_list)
+    list_parser = WOSListParser_v1_0_0(query_list=query_list)
     try:
         list_parser.parse()
     except ListQuerySyntaxError:
@@ -957,7 +959,7 @@ def test_list_parser_case_2() -> None:
 def test_list_parser_case_3() -> None:
     query_list = '1. TS=("Peer leader*" OR "Shared leader*")\n2. TS=("acrobatics" OR "acrobat" OR "acrobats")\n3. #1 AND #4\n'
 
-    list_parser = WOSListParser(query_list=query_list)
+    list_parser = WOSListParser_v1_0_0(query_list=query_list)
     try:
         list_parser.parse()
     except ListQuerySyntaxError as exc:
@@ -980,7 +982,7 @@ def test_list_parser_case_3() -> None:
 def test_list_parser_case_4() -> None:
     query_list = '1. TS=("Peer leader*" OR "Shared leader*" OR "Peer leader*" OR "Shared leader*")\n2. TS=("acrobatics" OR "acrobat" OR "acrobats")\n3. #1 AND #2\n'
     print(query_list)
-    list_parser = WOSListParser(query_list=query_list)
+    list_parser = WOSListParser_v1_0_0(query_list=query_list)
     query = list_parser.parse()
     print(query.to_string())
     assert (
@@ -1025,7 +1027,7 @@ def test_list_parser_case_4() -> None:
 def test_list_parser_case_5() -> None:
     query_list = '1. TS=("Peer leader*" OR "Shared leader*")\n2. TS=("acrobatics" OR "acrobat" OR "acrobats")\n3. #1 AND #2 AND\n'
 
-    list_parser = WOSListParser(query_list=query_list)
+    list_parser = WOSListParser_v1_0_0(query_list=query_list)
     try:
         list_parser.parse()
     except ListQuerySyntaxError as exc:
@@ -1050,7 +1052,7 @@ def test_list_parser_case_5() -> None:
 def test_list_parser_case_6() -> None:
     query_list = "1. TS=(inflammatory bowel diseases OR (inflamm* AND bowel*) OR (ulcer* colitis) OR crohn OR crohns OR ileitis or ileocolitis OR granulomatous enteritis OR proctocolitis OR regional enteritis OR rectosigmoiditis)\n2. TS=(prebiotic* OR synbiotic OR inulin OR galactan* or *oligosacc* OR pectin)\n3. TS=(interven* OR trial* or study)\n4. #3 AND #2 AND #1\n"
 
-    list_parser = WOSListParser(query_list=query_list)
+    list_parser = WOSListParser_v1_0_0(query_list=query_list)
     query = list_parser.parse()
     print(query.to_string())
     # Note: parentheses for inflamm* AND bowl* are missing?
@@ -1106,7 +1108,7 @@ def test_list_parser_case_6() -> None:
 def test_list_parser_case_7() -> None:
     query_list = "1. TS=(inflammatory bowel diseases OR ileitis or ileocolitis)\n2. TS=(prebiotic* OR synbiotic)\n3. #1 AND #2 AND PY=2000\n"
 
-    list_parser = WOSListParser(query_list=query_list)
+    list_parser = WOSListParser_v1_0_0(query_list=query_list)
     query = list_parser.parse()
     print(query.to_string())
     # Note: parentheses for inflamm* AND bowl* are missing?
@@ -1120,7 +1122,7 @@ def test_list_parser_case_7() -> None:
 def test_list_parser_case_8() -> None:
     query_list = '1. TS=("Peer leader*" OR "Shared leader*")\n2. TS=("acrobatics" OR "acrobat")\n3. #1 AND #2\n'
 
-    list_parser = WOSListParser(query_list=query_list)
+    list_parser = WOSListParser_v1_0_0(query_list=query_list)
     query = list_parser.parse()
     assert (
         query.to_string()
@@ -1129,7 +1131,7 @@ def test_list_parser_case_8() -> None:
 
     query_list = '1. "Peer leader*" OR "Shared leader*"\n2. "acrobatics" OR "acrobat"\n3. #1 AND #2\n'
 
-    list_parser = WOSListParser(query_list=query_list)
+    list_parser = WOSListParser_v1_0_0(query_list=query_list)
     query = list_parser.parse()
     print(query.to_structured_string())
     assert (
@@ -1215,7 +1217,7 @@ def test_parser(query_str: str, field_general: str, expected_parsed: str) -> Non
         f"Run query parser for: \n  {Colors.GREEN}{query_str}{Colors.END}\n--------------------\n"
     )
 
-    parser = WOSParser(query_str, field_general=field_general)
+    parser = WOSParser_v1_0_0(query_str, field_general=field_general)
     query = parser.parse()
 
     assert expected_parsed == query.to_generic_string(), print(
